+++ src/main/java/hu/nyirszikszi/vizsgaremek/cinema/dto/BookingResponse.java
@@ -5,9 +5,9 @@ import hu.nyirszikszi.vizsgaremek.cinema.enums.BookingStatus;
 import java.time.LocalDateTime;
 
 public record BookingResponse(
-        int bookingId,
+        Long bookingId,
         BookingStatus status,
         LocalDateTime created,
-        int showtimeId,
-        int seatId
+        Long showtimeId,
+        Long seatId
 ){}
