+++ src/main/java/hu/nyirszikszi/vizsgaremek/cinema/dto/CreateBookingRequest.java
@@ -1,6 +1,6 @@
 package hu.nyirszikszi.vizsgaremek.cinema.dto;
 
 public record CreateBookingRequest(
-        int showtimeId,
-        int SeatId
+        Long showtimeId,
+        Long SeatId
 ) {}
