+++ src/main/java/hu/nyirszikszi/vizsgaremek/cinema/dto/SeatAvailabilityResponse.java
@@ -1,9 +1,9 @@
 package hu.nyirszikszi.vizsgaremek.cinema.dto;
 
-import hu.nyirszikszi.vizsgaremek.cinema.enums.BookingStatus;
+
 
 public record SeatAvailabilityResponse(
-        int seatId,
+        Long seatId,
         int row,
         int number,
         String status
