+++ src/main/java/hu/nyirszikszi/vizsgaremek/cinema/repository/ShowtimeRepository.java
@@ -3,5 +3,5 @@ package hu.nyirszikszi.vizsgaremek.cinema.repository;
 import hu.nyirszikszi.vizsgaremek.cinema.entity.Showtime;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShowtimeRepository extends JpaRepository<Showtime, Integer> {
+public interface ShowtimeRepository extends JpaRepository<Showtime, Long> {
 }
