+++ src/main/java/hu/nyirszikszi/vizsgaremek/cinema/repository/TheaterRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface TheaterRepository extends JpaRepository<Theater,Integer> {
+public interface TheaterRepository extends JpaRepository<Theater,Long> {
 
     Optional<Theater> findByNameIgnoreCase(String name);
